+++ .atoum.php
@@ -13,8 +13,8 @@ use mageekguy\atoum;
 use mageekguy\atoum\reports;
 
 // Enable extension
-$extension = new reports\extension($script);
-$extension->addToRunner($runner);
+//$extension = new reports\extension($script);
+//$extension->addToRunner($runner);
 
 // Write all on stdout.
 $stdOutWriter = new atoum\writers\std\out();