+++ src/Repository/LanguageRepository.php
@@ -13,7 +13,7 @@ namespace App\Repository;
 
 use App\Entity\Language;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ManagerRegistry;
 
 class LanguageRepository extends ServiceEntityRepository
 {