+++ src/Repository/SearchRepository.php
@@ -13,7 +13,7 @@ namespace App\Repository;
 
 use App\Entity\Search;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ManagerRegistry;
 
 class SearchRepository extends ServiceEntityRepository
 {